# doc/openapi/dan.yaml
openapi: 3.0.0
info:
  title: DAN API
  description: >-
    ## Conventions


    The base URL to send all API requests is

    ```

    https://api.opendan.ai

    ```


    HTTPS is required for all API requests.


    The DAN API follows RESTful conventions when possible, with most operations
    performed via GET, POST, PATCH, and DELETE requests on page and database
    resources. Request and response bodies are encoded as JSON.


    At present, only HTTP requests are supported to complete image generation
    tasks.


    The APIs has similar respones.


    On success, dan server will response http `status code` 200 and an object in
    JSON as below:


    ```JSON

    {
        "code": 200,
        "data": `The returned object`,
        "message": "success"
    }

    ```


    On failure, dan server will response a non-ok http `status code` (i.e. not
    200, may be 401 for an example) and an object JSON as below:


    ```JSON

    {
        "code": `Error code`,
        "message": `A string describe the error`
    }

    ```


    ## Supported list values

    Under rapid expansion, please stay tuned.


    ### Models

    | Name                 |
    Value                                                            |
    Ref                                                                |
    Download                                                                                                        
    |

    | -------------------- |
...
    |

...
    |

...
    |

...




...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >-
    
        - Bearer DAN_API_KEY.

        - Public anonymous DAN_API_KEY is:

        ```

        sk-000000000000000000000000000000000000000000000000

        ```

        - Please refer to the [Authentication section](../Authentication.mdx) on how to obtain a private DAN_API_KEY.
...
paths:
  /api/sd/txt2img:
    post: &txt2img_post
      summary: /api/sd/txt2img
      deprecated: false
      description: This interface provides a service for generating images from text.
      tags:
        - Stable Diffusion API
      parameters: *common_parameters_basics
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                <<: *common_xxx2img_properties
                upscale:
                  type: object
                  properties:
                    denoising_strength:
                      type: number
                      description: >-
                        Controls the level of denoising; smaller values yield
                        results that are closer to the original generated image,
                        but may be blurry; larger values may lead the output
                        looks different from the original generated image and
                        may looks strange. Valid range is [0, 1], but I
                        recomment you make it between 0.4 and 0.6.
                    scale:
                      type: number
                      description: The upscale rate. Valid range is (1.0, 2.0].
                    upscaler:
                      type: string
                      description: >-
                        The upscaler algorithm name, Currently only Latent is
                        supported.
                  description: Optional, add it if you want to upscale the result.
              required:
                - prompt
                - sampler_name
                - width
                - height
                - model
            example:
              prompt: A cat
              sampler_name: DDIM
              width: 512
              height: 512
              model: 3a17d0deffa4592fd91c711a798031a258ab44041809ade8b4591c0225ea9401
              cfg_scale: 7
              negative_prompt: ''
              seed: -1
              loras:
                - - >-
                    62efe75048d55a096a238c6e8c4e12d61b36bf59e388a90589335f750923954c
                  - 0.7
              steps: 20
              upscale:
                denoising_strength: 0.5
                scale: 2
                upscaler: Latent
      responses:
        '200':
          <<: *common_xxx2img_200
          description: The generated image is returned here.
  /api/sd/txt2img/async:
    post:
      <<: *txt2img_post
      summary: /api/sd/txt2img/async
      description: >-
        This interface provides a service for generating images from text. <br/>
        Its usage is almost the same as `txt2img`. The difference between this interface and `txt2img` is that this interface will return immediately with task information.<br/>
        After obtaining the task information, you can use the [`/api/sd/task/status`](./api-sd-task-status.api.mdx) interface to query the task status and get the result.
      responses:
        '200':
          <<: *common_async_200
  /api/sd/img2img:
    post: &img2img_post
      summary: /api/sd/img2img
      deprecated: false
      description: This interface provides a service for generating images from images.
      tags:
        - Stable Diffusion API
      parameters: *common_parameters_basics
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                init_image:
                  type: string
                  description: The base64-encoded string of your original image
                denoising_strength:
                  type: number
                  description: "Controls the level of denoising; smaller values yield results that are closer to the original image. Valid range is [0, 1]"
                <<: *common_xxx2img_properties
                inpaint:
                  type: object
                  description: If you want to regenerate part of your image, this will be what you want.
                  properties:
                    mask:
                      type: string
                      description: This is a base64 encoded image of a mask which defined where you want to regenerate.
                    mask_blur:
                      type: number
                      description: Before regenerating, the image and mask will be gaussian blured by this radius. Valid range is [0, 64], default 0.
                    mask_mode:
                      type: number
                      description: 0 means regenerate where is masked; 1 means regenerate where is not masked.
                    inpaint_area:
                      type: number
                      description: 0 means regenerate whole image, then paste corresponding area back. 1 means only the masked area will be regenereate, then paste corresponding area back.
                  required:
                    - mask
              required:
                - prompt
                - sampler_name
                - width
                - height
                - denoising_strength
                - init_image
                - model
            example:
              prompt: A cat
              loras:
                - - >-
                    62efe75048d55a096a238c6e8c4e12d61b36bf59e388a90589335f750923954c
                  - 0.7
              seed: -1
              sampler_name: DDIM
              steps: 20
              cfg_scale: 7
              width: 512
              height: 512
              negative_prompt: ''
              model: 3a17d0deffa4592fd91c711a798031a258ab44041809ade8b4591c0225ea9401
              denoising_strength: 0.5
              init_image: >-
                iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAIAAACQkWg2AAAAwnRFWHRwYXJhbWV0ZXJzAEEgY2F0LDxsb3JhOmtvcmVhbkRvbGxMaWtlbmVzc192MTA6MC43PgpTdGVwczogMjAsIFNhbXBsZXI6IERESU0sIENGRyBzY2FsZTogNy4wLCBTZWVkOiAyMzAwMjY4MzQwLCBTaXplOiAxNngxNiwgTW9kZWwgaGFzaDogM2ExN2QwZGVmZiwgU2VlZCByZXNpemUgZnJvbTogLTF4LTEsIERlbm9pc2luZyBzdHJlbmd0aDogMOLvrY0AAAEWSURBVHicjZHBTkJBDEXPbWdAg4n+/3f5GSYuJIHHTFsXiKCCelYzyT1Nc6vXl+eHxyf3jbSGOxAfFAjy9E5ImKp6hQ4bvpNUIIGBoGAP3qCBIMC/Cobs4itYQzYA4sf4qzi4zf12LDs4/M+hbbc7804u/a6j9qdgy4hlHHIG85fFCgom0B42jyOGrNOvjv+MjmPLqqocb9ZWaH0Rmqfuj0JAgkCqKhjQT+mMHElkDbOJQqQhqMQER+HM5DBy5NxLU8yoxSy94Qgs4PI0H9vI0pzKOcahakQskXuRgE6HO9NZdQjX9KyxyAvMDMeg143iVw5OWU9wIyGgQYnVrSoFYaTokAbJLEy0q4Kgw71jYJAQRsAWeAf79pJ7zSui0wAAAABJRU5ErkJggg==
      responses:
        '200':
          <<: *common_xxx2img_200
          description: The generated image is returned here.
  /api/sd/img2img/async:
    post:
      <<: *img2img_post
      summary: /api/sd/img2img/async
      description: >-
        This interface provides a service for generating images from images. <br/>
        Its usage is almost the same as `img2img`. The difference between this interface and `img2img` is that this interface will return immediately with task information.<br/>
        After obtaining the task information, you can use the [`/api/sd/task/status`](./api-sd-task-status.api.mdx) interface to query the task status and get the result.
      responses:
        '200':
          <<: *common_async_200
  /api/sd/interrogate:
    post: &interrogate_post
      summary: /api/sd/interrogate
      deprecated: false
      description: 'Interrogate tasks will generate a description for an input image. '
      tags:
        - Stable Diffusion API
      parameters: *common_parameters_basics
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                image:
                  type: string
                  description: The base64-encoded string of your input image
                model:
                  type: string
                  description: >-
                    The model name used to describe the image, avalaible options
                    are clip and deepdanbooru
              required:
                - image
                - model
            example:
              image: >-
                iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAIAAACQkWg2AAAAwnRFWHRwYXJhbWV0ZXJzAEEgY2F0LDxsb3JhOmtvcmVhbkRvbGxMaWtlbmVzc192MTA6MC43PgpTdGVwczogMjAsIFNhbXBsZXI6IERESU0sIENGRyBzY2FsZTogNy4wLCBTZWVkOiAyMzAwMjY4MzQwLCBTaXplOiAxNngxNiwgTW9kZWwgaGFzaDogM2ExN2QwZGVmZiwgU2VlZCByZXNpemUgZnJvbTogLTF4LTEsIERlbm9pc2luZyBzdHJlbmd0aDogMOLvrY0AAAEWSURBVHicjZHBTkJBDEXPbWdAg4n+/3f5GSYuJIHHTFsXiKCCelYzyT1Nc6vXl+eHxyf3jbSGOxAfFAjy9E5ImKp6hQ4bvpNUIIGBoGAP3qCBIMC/Cobs4itYQzYA4sf4qzi4zf12LDs4/M+hbbc7804u/a6j9qdgy4hlHHIG85fFCgom0B42jyOGrNOvjv+MjmPLqqocb9ZWaH0Rmqfuj0JAgkCqKhjQT+mMHElkDbOJQqQhqMQER+HM5DBy5NxLU8yoxSy94Qgs4PI0H9vI0pzKOcahakQskXuRgE6HO9NZdQjX9KyxyAvMDMeg143iVw5OWU9wIyGgQYnVrSoFYaTokAbJLEy0q4Kgw71jYJAQRsAWeAf79pJ7zSui0wAAAABJRU5ErkJggg==
              model: clip
      responses:
        '200':
          <<: *common_interrogate_200
          description: The input image's description is returned here.
  /api/sd/interrogate/async:
    post:
      <<: *interrogate_post
      summary: /api/sd/interrogate/async
      deprecated: false
      description: >-
        Interrogate tasks will generate a description for an input image.<br/>
        Its usage is almost the same as `interrogate`. The difference between this interface and `interrogate` is that this interface will return immediately with task information.<br/>
        After obtaining the task information, you can use the [`/api/sd/task/status`](./api-sd-task-status.api.mdx) interface to query the task status and get the result.
      tags:
        - Stable Diffusion API
      responses:
        '200':
          <<: *common_async_200
  /api/sd/upscale:
    post: &upscale_post
      summary: /api/sd/upscale
      deprecated: false
      description: Upscale the input image.
      tags:
        - Stable Diffusion API
      parameters: *common_parameters_basics
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                image:
                  type: string
                  description: The base64-encoded string of your input image
                upscalingRate:
                  type: number
                  description: >-
                    How many times you want the image be scaled,
                    the output image's width and height will be multiplied by this rate. Valid range is (1, 10].
                    If you set this rate to large, the task may fail due to memory issue.
                upscaler:
                  type: string
                  description: >-
                    The upscaling algorithm to use. Valid values are listed below:
                    <table>
                      <tr>
                        <th>Option</th>
                        <th>Description</th>
                      </tr>
                      <tr>
                        <td>Lanczos</td>
                        <td>This is a mathematical function-based interpolation method that can keep the edges of the image clear and avoid the jagged effect. It is suitable for video enlargement, but for static images, it may produce ringing or oversharpening effects.</td>
                      </tr>
                      <tr>
                        <td>Neareast</td>
                        <td>This is the simplest interpolation method that just copies each pixel in the image to adjacent positions, enlarging the size of the image by 3. Its advantage is that it is fast, but the disadvantage is that the enlarged image will be very blurry and distorted.</td>
                      </tr>
                      <tr>
                        <td>ESRGAN 4x</td>
                        <td>This is a deep learning based super-resolution method that uses an Enhanced Super Resolution Generative Adversarial Network to enlarge the image by 4. Its advantage is that it can restore details and textures in the image, but the disadvantage is that it requires large computing resources and training data.</td>
                      </tr>
                      <tr>
                        <td><del>LDSR</del></td>
                        <td>Disabled since it's very expensive and it produces poor outputs. This is an image generation method based on stable diffusion that can enlarge the image while maintaining the style and texture of the image. Its advantage is that it can handle various types of images, but the disadvantage is that it requires a long time and a lot of memory.</td>
                      </tr>
                      <tr>
                        <td>R-ESRGAN 4x+</td>
                        <td>This is an improved method based on Real-ESRGAN, which uses residual connections and adaptive instance normalization to improve the clarity and naturalness of the image. Its advantage is that it can restore details and textures in the image, but the disadvantage is that it requires large computing resources and training data.</td>
                      </tr>
                      <tr>
                        <td>R-ESRGAN 4x+ Anime6B</td>
                        <td>This is an image enlargement method specifically designed for anime style. It uses the R-ESRGAN 4x+ model and is fine-tuned on the Anime6B dataset. Its advantage is that it can maintain the color and lines of anime images, but the disadvantage is that it may not be suitable for other types of images.</td>
                      </tr>
                      <tr>
                        <td>SwinIR 4x</td>
                        <td>This is an image restoration method based on Swin Transformer, which uses self-attention mechanism and multi-scale fusion to enlarge the image. Its advantage is that it can adapt to different scenes and noise levels, but the disadvantage is that it requires high hardware configuration and memory consumption.</td>
                      </tr>
                    </table>
                upscaler2:
                  type: string
                  description: >-
                    Required only if `upscaler2_factor` > 0.
                    The upscaling algorithm to use. Valid values are same as `upscaler`.
                    This field is necessary only if `upscaler2_factor` > 0, if so, denoting `I` as the input image, 
                    the final image will be: `blend(upscaler(I), upscaler2(I), upscaler2_factor)`,
                    i.e. `upscaler(I) * (1-upscaler2_factor) + upscaler2(I) * upscaler2_factor`.
                upscaler2_factor:
                  type: number
                  description: See `upscaler2`. Valid range is [0, 1), default 0.
              required:
                - image
                - upscalingRate
                - upscaler
            example:
              image: >-
                iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAIAAACQkWg2AAAAwnRFWHRwYXJhbWV0ZXJzAEEgY2F0LDxsb3JhOmtvcmVhbkRvbGxMaWtlbmVzc192MTA6MC43PgpTdGVwczogMjAsIFNhbXBsZXI6IERESU0sIENGRyBzY2FsZTogNy4wLCBTZWVkOiAyMzAwMjY4MzQwLCBTaXplOiAxNngxNiwgTW9kZWwgaGFzaDogM2ExN2QwZGVmZiwgU2VlZCByZXNpemUgZnJvbTogLTF4LTEsIERlbm9pc2luZyBzdHJlbmd0aDogMOLvrY0AAAEWSURBVHicjZHBTkJBDEXPbWdAg4n+/3f5GSYuJIHHTFsXiKCCelYzyT1Nc6vXl+eHxyf3jbSGOxAfFAjy9E5ImKp6hQ4bvpNUIIGBoGAP3qCBIMC/Cobs4itYQzYA4sf4qzi4zf12LDs4/M+hbbc7804u/a6j9qdgy4hlHHIG85fFCgom0B42jyOGrNOvjv+MjmPLqqocb9ZWaH0Rmqfuj0JAgkCqKhjQT+mMHElkDbOJQqQhqMQER+HM5DBy5NxLU8yoxSy94Qgs4PI0H9vI0pzKOcahakQskXuRgE6HO9NZdQjX9KyxyAvMDMeg143iVw5OWU9wIyGgQYnVrSoFYaTokAbJLEy0q4Kgw71jYJAQRsAWeAf79pJ7zSui0wAAAABJRU5ErkJggg==
              upscalingRate: 2
              upscaler: Neareast
      responses:
        '200':
          <<: *common_xxx2img_200
          description: The upscaled image is returned here.
  /api/sd/upscale/async:
    post:
      <<: *upscale_post
      summary: /api/sd/upscale/async
      deprecated: false
      description: >-
        Upscale the input image.<br/>
        Its usage is almost the same as `upscale`. The difference between this interface and `upscale` is that this interface will return immediately with task information.<br/>
        After obtaining the task information, you can use the [`/api/sd/task/status`](./api-sd-task-status.api.mdx) interface to query the task status and get the result.
      tags:
        - Stable Diffusion API
      responses:
        '200':
          <<: *common_async_200
  /api/sd/task/status:
    post:
      summary: /api/sd/task/status
      deprecated: false
      description: This API is used to query task's progress or retrive the task's result.
      tags:
        - Task
      parameters: *common_parameters_basics
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                taskId:
                  type: string
                  description: The task id returned by asynchronous API, which is the task you care about.
              required:
                - taskId
            example:
              image: >-
                iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAIAAACQkWg2AAAAwnRFWHRwYXJhbWV0ZXJzAEEgY2F0LDxsb3JhOmtvcmVhbkRvbGxMaWtlbmVzc192MTA6MC43PgpTdGVwczogMjAsIFNhbXBsZXI6IERESU0sIENGRyBzY2FsZTogNy4wLCBTZWVkOiAyMzAwMjY4MzQwLCBTaXplOiAxNngxNiwgTW9kZWwgaGFzaDogM2ExN2QwZGVmZiwgU2VlZCByZXNpemUgZnJvbTogLTF4LTEsIERlbm9pc2luZyBzdHJlbmd0aDogMOLvrY0AAAEWSURBVHicjZHBTkJBDEXPbWdAg4n+/3f5GSYuJIHHTFsXiKCCelYzyT1Nc6vXl+eHxyf3jbSGOxAfFAjy9E5ImKp6hQ4bvpNUIIGBoGAP3qCBIMC/Cobs4itYQzYA4sf4qzi4zf12LDs4/M+hbbc7804u/a6j9qdgy4hlHHIG85fFCgom0B42jyOGrNOvjv+MjmPLqqocb9ZWaH0Rmqfuj0JAgkCqKhjQT+mMHElkDbOJQqQhqMQER+HM5DBy5NxLU8yoxSy94Qgs4PI0H9vI0pzKOcahakQskXuRgE6HO9NZdQjX9KyxyAvMDMeg143iVw5OWU9wIyGgQYnVrSoFYaTokAbJLEy0q4Kgw71jYJAQRsAWeAf79pJ7zSui0wAAAABJRU5ErkJggg==
              model: clip
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  data:
                    type: object
                    properties:
                      caption:
                        type: string
                        description: Maybe `undefined`. If the task a an interrogate task, This is the input image description.
                      images:
                        type: array
                        items:
                          type: string
                        description: >-
                          May be `undefined`. If the task is txt2img/img2img task, this is the generated file list, encoded in base64.
                      taskId:
                        type: string
                      queuePosition:
                        type: integer
                        description: >-
                          The length of the queue of pending tasks, when status
                          != 0, queuePosition = 0
                      status:
                        type: integer
                        description: 0:pending, 1:processing, 2:success, 3:fail
                    required:
                      - taskId
                      - queuePosition
                      - status
                  message:
                    type: string
                required:
                  - code
                  - data
                  - message
              examples:
                Success:
                  summary: Success
                  value:
                    code: 200
                    data:
                      taskId: faa37545-b284-461d-85fc-32468c7af036
                      queuePosition: 0
                      status: 2
                      caption: >-
                        a blurry image of a white and blue background with a
                        blurry effect of a blue and white background, Carlos
                        Trillo Name, behance hd, a detailed matte painting, net
                        art
                    message: success
          description: >-
            If the task has not been finished, only the task information is returned.
            Else return the task output.
servers:
...